refactor: redistributed definition of struct DeskUpIsAvailable into a separate file, and minor changes regarding this
