+++ source/desk_up_window_backend/desk_up_is_available.h
@@ -1,0 +1,9 @@
+#ifndef DESKUPISAVAILABLE_H
+#define DESKUPISAVAILABLE_H
+
+struct DeskUpisAvailable{
+    const char * name;
+    bool (*isAvailable)(void);
+};
+
+#endif
