+++ source/desk_up_window_backend/window_backends/desk_up_win/desk_up_win.h
@@ -10,6 +10,7 @@
 #include "window_desc.h"
 #include "desk_up_window_bootstrap.h"
 #include "desk_up_window_device.h"
+#include "desk_up_is_available.h"
 
 /**
  * This struct is an opaque type used to pass platform-specific info from a window to the backend calls. 
@@ -17,11 +18,18 @@
  */
 struct windowData;
 
+extern std::unique_ptr<HWND> desk_up_hwnd;
+
 DeskUpWindowDevice * WIN_CreateDevice(HWND deskUpHWND);
 
 extern DeskUpWindowBootStrap winWindowDevice;
 
+extern DeskUpisAvailable winIsAvailable;
+
+bool WIN_isAvailable();
+
 std::string WIN_getDeskUpPath();
+
 unsigned int WIN_getWindowHeight(DeskUpWindowDevice * _this);
 unsigned int WIN_getWindowWidth(DeskUpWindowDevice * _this);
 unsigned int WIN_getWindowXPos(DeskUpWindowDevice * _this);
