+++ source/desk_up_window_backend/window_core.h
@@ -10,15 +10,7 @@
 
 extern std::string DESKUPDIR;
 
-struct DeskUpisAvailable;
-
-bool WIN_isAvailable();
-
-extern DeskUpisAvailable win;
-
 extern DeskUpWindowDevice * current_window_backend;
-
-extern HWND desk_up_hwnd;
 
 //function to initialize the backend and choose the correct device. Previously used x11 and Windows, but now only connects windows
 int DU_Init(HWND thisHwnd);
